pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.45",
    "boto3-stubs>=1.42.48",
    "fastapi>=0.128.6",
    "mangum>=0.21.0",
    "pendulum>=3.2.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.0",
    "ujson>=5.11.0",
    "uvicorn>=0.40.0",
    "argon2>=0.1.10",
    "pyjwt>=2.11.0",
]

[dependency-groups]
dev = [
    "aws-lambda-powertools[validation]>=3.24.0",
    "bandit>=1.9.3",
    "moto>=5.1.21",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.2.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.0",
    "time-machine>=3.2.0",
    "ty>=0.0.15",
]

[tool.pytest.ini_options]
addopts = "--cache-clear --cov-branch --cov-report term --cov=app/ -n 4 -r f"
env = [
    "APP_NAME=user-service",
    "AWS_ACCESS_KEY_ID=access_key_id",
    "AWS_DEFAULT_REGION=eu-central-1",
    "AWS_SECRET_ACCESS_KEY=secret_access_key",
    "DEBUG=true",
    "DEFAULT_TIMEZONE=UTC",
    "LOG_LEVEL=INFO",
    "POWERTOOLS_LOGGER_LOG_EVENT=true",
    "POWERTOOLS_SERVICE_NAME=user-service",
    "POWERTOOLS_DEBUG=false",
    "STAGE=test",
]
pythonpath = [
    ".",
    "app",
]
testpaths = [
    "tests"
]
